build: registration refactor
